CLI integration tests: Set reliable default keyserver.
In Tails, the default keyserver is sometimes flaky [0], so
we implemented a retry in securedrop-admin update [1].

In the integration tests, using the same default pool as in
Tails, there appears to still be some flakiness. Digging a bit
my hypothesis is that this is occuring when we are assigned a keyserver
that is IPv6 from the pool [2]. For test purposes, I'm setting a
reliable default keyserver in gpg.conf.

[0] https://labs.riseup.net/code/issues/12689
[1] https://github.com/freedomofpress/securedrop/pull/3257
[2] https://github.com/nodejs/docker-node/issues/380#issuecomment-306451705