Links out to Trusty install docs in Qubes docs
We already have copious documentation on how to install Trusty from an
ISO, so let's link out to that to be sure. Tacked on a reference in the
Trusty docs so that we can link directly to the install part.

Removes the recommendation to edit /etc/hosts on sd-dev, since we can
trust the provisioning scripts to manage access to the VMs for us.